refactor : project 생성 api 서비스 로직 분리
목적 : Project 서비스의 인지 복잡도가 25로 매우 높으므로 서비스 로직을 분리함
