+++ sequence_member/src/main/java/sequence/sequence_member/project/controller/ProjectController.java
@@ -20,6 +20,7 @@ import sequence.sequence_member.project.dto.ProjectInputDTO;
 import sequence.sequence_member.project.dto.ProjectOutputDTO;
 import sequence.sequence_member.project.dto.ProjectUpdateDTO;
 import sequence.sequence_member.project.service.ProjectBookmarkService;
+import sequence.sequence_member.project.service.ProjectCreateService;
 import sequence.sequence_member.project.service.ProjectService;
 
 import java.util.ArrayList;
@@ -32,12 +33,13 @@ import java.util.List;
 public class ProjectController {
 
     private final ProjectService projectService;
+    private final ProjectCreateService projectCreateService;
     private final ProjectBookmarkService projectBookmarkService;
 
     @PostMapping()
     public ResponseEntity<ApiResponseData<String>> registerProject(@Valid @RequestBody ProjectInputDTO projectInputDTO, @AuthenticationPrincipal
                                                                    CustomUserDetails customUserDetails) {
-        projectService.createProject(projectInputDTO, customUserDetails.getUsername());
+        projectCreateService.createProject(projectInputDTO, customUserDetails.getUsername());
         return ResponseEntity.ok(ApiResponseData.success(null, "프로젝트 등록 성공"));
     }
